requirements.txt
# Package name = version specifier

annotated-types==0.6.0
anyio==3.7.1
boto3==1.29.6
botocore==1.32.6
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
fastapi==0.104.1
fonttools==4.45.1
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pdf2image==1.16.3
Pillow==10.1.0
pip==23.2.1
pydantic==2.5.2
pydantic_core==2.14.5
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
pyzbar==0.1.9
requests==2.31.0
s3transfer==0.7.0
setuptools==65.5.0
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
zipp==3.17.0
uvicorn==0.24.0.post1